url/data-uri-fragment.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Data URI parsing of fragments</title>
<link rel="help" href="https://url.spec.whatwg.org/">
<meta name="assert" content="Fragments should not be included as part of a data URI's body">
<link rel="match" href="data-uri-fragment-ref.html">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<iframe id="iframe"></iframe>

<script>
const TEST_JS = `data:text/html,
    <style>:target { color: green; }<\/style>
    <script>window.addEventListener('load', function() {
      const data = {
        foo_matches_target_selector: document.getElementById('foo').matches(':target'),
        body_html: document.body.innerHTML,
      };
      parent.postMessage(data, '*');
    });<\/script>
    <p id="foo">This should be the only visible text.</p>#foo`.replace('\n', '');

async_test(function(t) {
  window.addEventListener("message", t.step_func_done(function(event) {
    console.log(event);
    assert_true(event.data.foo_matches_target_selector);
    assert_equals(event.data.body_html,
        '<p id="foo">This should be the only visible text.</p>');
  }));

  const iframe = document.getElementById("iframe");
  iframe.src = TEST_JS;
});
</script>
