third_party/WebKit/LayoutTests/external/wpt/url/data-uri-fragment-expected.txt

This is a testharness.js-based test.
FAIL Data URI parsing of fragments assert_equals: expected "<p id=\"foo\">This should be the only visible text.</p>" but got "<p id=\"foo\">This should be the only visible text.</p>#foo"
Harness: the test ran to completion.
